Updated guide to incorporate 'start' and 'complete' sections.
I had the guide working, but needed to properly have all the 'start' sections
first and then transition suitably to the 'complete' state. This
included adjusting some of the 'start' code and removing things that didn't
make sense in that initial cut of stuff. I also trimmed down the user data
since we didn't need all the entries.